script to identify the "connected" TfL bus routes
for every bus route pair, it checks for the last 3 stops of a line and the first
3 stops of the other, if any stop is within 500 meters of another, that line is
called "connected" (as they are pretty close to each other).

we chose just to compare start/end of lines to invite to use the whole line, not
just one part and then jump on another route.